Moved around socials and did some minor styling changes.
