+++ src/components/about/SelfPortrait.tsx
@@ -1,5 +1,5 @@
 import portrait from '../../assets/images/portrait.avif'
 
-const SelfPortrait = () => (<img src={portrait} alt='Steven rock climbing' className='hover:grayscale'/>)
+const SelfPortrait = () => (<img src={portrait} alt='Steven rock climbing' className='hover:grayscale hover:scale-110'/>)
 
 export default SelfPortrait;
