+++ src/index.css
@@ -54,4 +54,24 @@
 	.card-toggle {
 		transform: rotateX(180deg);
 	}
+
+	.highlight {
+		background: linear-gradient(
+			to right,
+			oklch(var(--s)),
+			oklch(var(--a)),
+			oklch(var(--p)),
+			oklch(var(--a))
+		);
+		-webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    background-size: 500% auto;
+		animation: highlight 5s ease-in-out infinite alternate;
+	}
+
+	@keyframes highlight {
+		0% { background-position: 0 50%; }
+		100% { background-position: 100% 50%; }
+	}
 }
